requirements.txt
numpy
opencv-python
tqdm
matplotlib
seaborn
Pillow
torch
torchvision
